clean mdbt
