--- tatk/dst/mdbt/__init__.py
@@ -1,1 +1,0 @@
-from tatk.dst.mdbt.mdbt import MDBT
